+++ src/dogfight_remake/entities/planes/ww/WW1.java
@@ -190,7 +190,7 @@ public class WW1 extends Entity {
 	 * Paint method
 	 */
 	@Override
-	public void render(GameContainer container, Graphics g, int delta) {
+	public void render(GameContainer container, Graphics g, float delta) {
 		if (broken) {
 			return;
 		}
